+++ Network/ProtoBSDIP.h
@@ -35,7 +35,7 @@
 #include <winsock.h>
 #else
 // (that should just about cover it... --ryan.)
-#include <stdio.h>
+#include <cstdio>
 #include <unistd.h>
 #include <sys/types.h>
 #include <sys/socket.h>
@@ -91,7 +91,7 @@ class RProtocolBSDIP : public RSocket::RProtocol
       struct AddressIP {
          public:
             RSocket::ProtoType   prototype;        // Type of protocol
-            long                 lAddressLen;      // Actual address length (always <= MaxAddressSize)
+            int32_t                 lAddressLen;      // Actual address length (always <= MaxAddressSize)
             SOCKADDR_IN          address;          // Address
 
             bool operator==(const AddressIP& rhs) const
@@ -181,27 +181,27 @@ class RProtocolBSDIP : public RSocket::RProtocol
       // Send data - only valid with connected sockets
       virtual int16_t Send(                             // Returns 0 if data was sent
          void * pBuf,                                 // In:  Pointer to data buffer
-         long lNumBytes,                              // In:  Number of bytes to send
-         long *plActualBytes);                        // Out: Acutal number of bytes sent
+         int32_t lNumBytes,                              // In:  Number of bytes to send
+         int32_t *plActualBytes);                        // Out: Acutal number of bytes sent
 
       // SendTo - send data to specified address - for unconnected sockets
       virtual int16_t SendTo(                           // Returns 0 if data was sent
          void* pBuf,                                  // In:  Pointer to data buffer
-         long lNumBytes,                              // In:  Number of bytes to send
-         long* plActualBytes,                         // Out: Actual number of bytes sent
+         int32_t lNumBytes,                              // In:  Number of bytes to send
+         int32_t* plActualBytes,                         // Out: Actual number of bytes sent
          RSocket::Address* paddress);                 // In:  Address to send to
 
       // Receive data - only valid for connected sockets
       virtual int16_t Receive(                          // Returns 0 if data was received
          void* pBuf,                                  // In:  Pointer to data buffer
-         long lMaxBytes,                              // In:  Maximum number of bytes that fit in the buffer
-         long* plActualBytes);                        // Out: Actual number of bytes received into buffer
+         int32_t lMaxBytes,                              // In:  Maximum number of bytes that fit in the buffer
+         int32_t* plActualBytes);                        // Out: Actual number of bytes received into buffer
 
       // RecieveFrom - receive data from given address
       virtual int16_t ReceiveFrom(                      // Returns 0 if data was received
          void* pBuf,                                  // In:  Pointer to data buffer
-         long lMaxBytes,                              // In:  Maximum bytes that can fit in buffer
-         long* plActualBytes,                         // Out:  Actual number of bytes received into buffer
+         int32_t lMaxBytes,                              // In:  Maximum bytes that can fit in buffer
+         int32_t* plActualBytes,                         // Out:  Actual number of bytes received into buffer
          RSocket::Address* paddress);                 // Out: Source address returned here
 
       // Check if connection can be accepted without blocking
@@ -214,7 +214,7 @@ class RProtocolBSDIP : public RSocket::RProtocol
       virtual bool CanReceiveWithoutBlocking(void);
 
       // See how much data can be received without blocking
-      virtual long CheckReceivableBytes(void);
+      virtual int32_t CheckReceivableBytes(void);
 
       // Report error status
       virtual bool IsError(void);
@@ -234,11 +234,11 @@ class RProtocolBSDIP : public RSocket::RProtocol
 
       // Get the maximum datagram size supported by this socket
       static int16_t GetMaxDatagramSize(                // Returns 0 if info is available
-         long* plSize);                               // Out: Maximum datagram size (in bytes)
+         int32_t* plSize);                               // Out: Maximum datagram size (in bytes)
 
       // Get maximum number of sockets
       static int16_t GetMaxSockets(                     // Returns 0 if successfull, non-zero otherwise
-         long* plNum);
+         int32_t* plNum);
 
       // Get address from the socket
       static int16_t GetAddress(                        // Returns 0 if successfull, non-zero otherwise
