+++ Thing/GoalTimer.h
@@ -51,7 +51,7 @@ class CGoalTimer : public CThing
       double m_dZ;                                    // z coord
       RImage* m_pImage;                               // Pointer to only image (replace with 3d anim, soon)
       CSprite2 m_sprite;                              // Sprite (replace with CSprite3, soon)
-      long     m_lTimerMS;                            // Timer value in Milliseconds
+      int32_t     m_lTimerMS;                            // Timer value in Milliseconds
       int16_t    m_sKillGoal;                           // Number of kills to achieve
       int16_t    m_sUpDown;                             // Up or down timer
 
@@ -116,7 +116,7 @@ class CGoalTimer : public CThing
          RFile* pFile,                                // In:  File to load from
          bool bEditMode,                              // In:  True for edit mode, false otherwise
          int16_t sFileCount,                            // In:  File count (unique per file, never 0)
-         ULONG ulFileVersion);                        // In:  Version of file format to load.
+         uint32_t ulFileVersion);                        // In:  Version of file format to load.
 
       // Save object (should call base class version!)
       int16_t Save(                                     // Returns 0 if successfull, non-zero otherwise
