+++ Thing/Ladder.h
@@ -34,7 +34,7 @@
 
 #include <RSPiX.h>
 #include <Realm.h>
-#include <Character.h>
+#include <Thing/Thing3D/Character.h>
 
 class CLadder : public CThing
    {
@@ -140,7 +140,7 @@ class CLadder : public CThing
          RFile* pFile,                                // In:  File to load from
          bool bEditMode,                              // In:  True for edit mode, false otherwise
          int16_t sFileCount,                            // In:  File count (unique per file, never 0)
-         ULONG ulFileVersion);                        // In:  Version of file format to load.
+         uint32_t ulFileVersion);                        // In:  Version of file format to load.
 
       // Save object (should call base class version!)
       int16_t Save(                                     // Returns 0 if successfull, non-zero otherwise
